fix: createVote
